#4523 - Factor endpoint-specific FTS code out into adapter classes
- Further disentangle test code
