+++ inception/inception-kb/src/test/java/de/tudarmstadt/ukp/inception/kb/querybuilder/VirtuosoRepositoryTest.java
@@ -28,11 +28,9 @@ import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Method;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.eclipse.rdf4j.model.impl.SimpleValueFactory;
 import org.eclipse.rdf4j.repository.Repository;
-import org.eclipse.rdf4j.repository.RepositoryConnection;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
@@ -49,7 +47,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 
 import de.tudarmstadt.ukp.inception.kb.RepositoryType;
 import de.tudarmstadt.ukp.inception.kb.model.KnowledgeBase;
-import de.tudarmstadt.ukp.inception.kb.querybuilder.SPARQLQueryBuilderTest.Scenario;
+import de.tudarmstadt.ukp.inception.kb.querybuilder.SPARQLQueryBuilderLocalTestScenarios.Scenario;
 
 // The repository has to exist and Virtuoso has to be running for this test to run.
 // To create the repository, use the following steps:
@@ -58,23 +56,26 @@ import de.tudarmstadt.ukp.inception.kb.querybuilder.SPARQLQueryBuilderTest.Scena
 // Go to http://localhost:8890/conductor -> System Admin -> User Accounts -> SPARQL and add account role SPARQL_UPDATE
 // Check if FTS is enabled: `SELECT * from DB.DBA.RDF_OBJ_FT_RULES;`
 // Enable FTS: `DB.DBA.RDF_OBJ_FT_RULE_ADD (null, null, 'ALL');`
-@Disabled("not sure if this still works ok")
+@Disabled("It seems the FTS is not properly flushing its data and tests may fail kind of randomly")
 @Testcontainers(disabledWithoutDocker = true)
 public class VirtuosoRepositoryTest
 {
+    private static final String GRAPH_IRI = "http://testgraph";
+
     private static final Logger LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
     private static final int VIRTUOSO_PORT = 8890;
 
     @Container
     private static final GenericContainer<?> VIRTUOSO = new GenericContainer<>(
-            "openlink/virtuoso-opensource-7") //
+            "openlink/virtuoso-opensource-7:latest") //
                     .withEnv("DBA_PASSWORD", "secret") //
                     .withExposedPorts(VIRTUOSO_PORT) //
                     .waitingFor(Wait.forHttp("/sparql").forPort(VIRTUOSO_PORT)
                             .withStartupTimeout(ofMinutes(2)));
 
     private Repository repository;
+
     private KnowledgeBase kb;
 
     @BeforeEach
@@ -84,38 +85,38 @@ public class VirtuosoRepositoryTest
         System.out.printf("\n=== %s === %s =====================\n", methodName,
                 aTestInfo.getDisplayName());
 
-        suspendSslVerification();
-
         assertThat(VIRTUOSO.isRunning()).isTrue();
+
         virtuosoSqlCommand("DB.DBA.RDF_OBJ_FT_RULE_ADD (null, null, 'ALL');");
+        virtuosoSqlCommand("DB.DBA.VT_BATCH_UPDATE ('DB.DBA.RDF_OBJ', 'OFF', null);");
+
+        suspendSslVerification();
 
         kb = new KnowledgeBase();
         kb.setDefaultLanguage("en");
-        kb.setDefaultDatasetIri("http://testgraph");
+        kb.setDefaultDatasetIri(GRAPH_IRI);
         kb.setType(RepositoryType.REMOTE);
         kb.setFullTextSearchIri(FTS_VIRTUOSO.stringValue());
         kb.setMaxResults(100);
 
-        SPARQLQueryBuilderTest.initRdfsMapping(kb);
+        SPARQLQueryBuilderLocalTestScenarios.initRdfsMapping(kb);
 
-        repository = SPARQLQueryBuilderTest
+        repository = SPARQLQueryBuilderLocalTestScenarios
                 .buildSparqlRepository("http://dba:secret@" + VIRTUOSO.getHost() + ":"
                         + VIRTUOSO.getMappedPort(VIRTUOSO_PORT) + "/sparql-auth/");
 
-        // repository = SPARQLQueryBuilderTest
-        // .buildSparqlRepository("http://dba:secret@localhost:8890/sparql-auth/");
-
-        try (RepositoryConnection conn = repository.getConnection()) {
+        try (var conn = repository.getConnection()) {
             var ctx = SimpleValueFactory.getInstance().createIRI(kb.getDefaultDatasetIri());
             conn.clear(ctx);
         }
     }
 
-    private void virtuosoSqlCommand(String command) throws IOException, InterruptedException
+    static void virtuosoSqlCommand(String command) throws IOException, InterruptedException
     {
+        command = command.replace("\"", "\\\"");
         LOG.info("Running virtuoso command: {}", command);
-        var result = VIRTUOSO.execInContainer("echo", command, "|", "isql", "-U", "dba", "-P",
-                "secret");
+        var result = VIRTUOSO.execInContainer("sh", "-c",
+                "echo \"" + command + "\" | isql -U dba -P secret");
         LOG.info("Command exit code: {}", result.getExitCode());
         LOG.info("Command stdout: {}", result.getStdout());
         LOG.info("Command stderr: {}", result.getStderr());
@@ -141,10 +142,10 @@ public class VirtuosoRepositoryTest
                 // Virtuoso does not like to import the test data for this one
                 "testWithLabelStartingWith_OLIA");
 
-        return SPARQLQueryBuilderTest.tests().stream() //
-                .filter(scenario -> !exclusions.contains(scenario.name))
-                .map(scenario -> Arguments.of(scenario.name, scenario))
-                .collect(Collectors.toList());
+        return SPARQLQueryBuilderLocalTestScenarios.tests().stream() //
+                .filter(scenario -> !exclusions.contains(scenario.name)) //
+                .map(scenario -> Arguments.of(scenario.name, scenario)) //
+                .toList();
     }
 
     @ParameterizedTest(name = "{index}: test {0}")
